+++ src/slices/Signup.ts
@@ -13,8 +13,7 @@ export type NextPageTypes =
   | "SIGNUP_ENTER_CODE"
   | "PROCESS_EMAIL_CODE"
   | "SIGNUP_CREDENTIALS_ERROR"
-  | "SIGNUP_CREDENTIAL_PASSWORD"
-  | "SIGNUP_MFA"
+  | "SIGNUP_CREDENTIALS"
   | "SIGNUP_USER_CREATED"
   | "SIGNUP_CONFIRM_PASSWORD";
